+++ alpha-api/src/main/java/at/ac/tuwien/kr/alpha/api/programs/ASPCore2Program.java
@@ -1,7 +1,7 @@
 package at.ac.tuwien.kr.alpha.api.programs;
 
-import at.ac.tuwien.kr.alpha.api.rules.Head;
 import at.ac.tuwien.kr.alpha.api.rules.Rule;
+import at.ac.tuwien.kr.alpha.api.rules.heads.Head;
 
 public interface ASPCore2Program extends Program<Rule<Head>> {
