+++ alpha-api/src/main/java/at/ac/tuwien/kr/alpha/api/programs/Program.java
@@ -3,8 +3,8 @@ package at.ac.tuwien.kr.alpha.api.programs;
 import java.util.List;
 
 import at.ac.tuwien.kr.alpha.api.programs.atoms.Atom;
-import at.ac.tuwien.kr.alpha.api.rules.Head;
 import at.ac.tuwien.kr.alpha.api.rules.Rule;
+import at.ac.tuwien.kr.alpha.api.rules.heads.Head;
 
 public interface Program<R extends Rule<? extends Head>> {
