--- alpha-api/src/main/java/at/ac/tuwien/kr/alpha/api/rules/Head.java
@@ -1,4 +1,0 @@
-package at.ac.tuwien.kr.alpha.api.rules;
-
-public interface Head {
-}
