+++ alpha-api/src/main/java/at/ac/tuwien/kr/alpha/api/rules/NormalRule.java
@@ -1,5 +1,7 @@
 package at.ac.tuwien.kr.alpha.api.rules;
 
+import at.ac.tuwien.kr.alpha.api.rules.heads.NormalHead;
+
 public interface NormalRule extends Rule<NormalHead> {
 
 }
