+++ alpha-api/src/main/java/at/ac/tuwien/kr/alpha/api/rules/heads/DisjunctiveHead.java
@@ -1,4 +1,4 @@
-package at.ac.tuwien.kr.alpha.api.rules;
+package at.ac.tuwien.kr.alpha.api.rules.heads;
 
 public interface DisjunctiveHead extends Head{
