+++ alpha-api/src/main/java/at/ac/tuwien/kr/alpha/api/rules/heads/Head.java
@@ -1,0 +1,4 @@
+package at.ac.tuwien.kr.alpha.api.rules.heads;
+
+public interface Head {
+}
