+++ alpha-api/src/main/java/at/ac/tuwien/kr/alpha/api/rules/heads/NormalHead.java
@@ -1,4 +1,4 @@
-package at.ac.tuwien.kr.alpha.api.rules;
+package at.ac.tuwien.kr.alpha.api.rules.heads;
 
 import at.ac.tuwien.kr.alpha.api.programs.atoms.Atom;
