+++ alpha-core/src/main/java/at/ac/tuwien/kr/alpha/core/programs/AbstractProgram.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 import at.ac.tuwien.kr.alpha.api.programs.InlineDirectives;
 import at.ac.tuwien.kr.alpha.api.programs.atoms.Atom;
-import at.ac.tuwien.kr.alpha.api.rules.Head;
 import at.ac.tuwien.kr.alpha.api.rules.Rule;
+import at.ac.tuwien.kr.alpha.api.rules.heads.Head;
 import at.ac.tuwien.kr.alpha.commons.util.Util;
 
 /**
