+++ alpha-core/src/main/java/at/ac/tuwien/kr/alpha/core/rules/CompiledRule.java
@@ -3,8 +3,8 @@ package at.ac.tuwien.kr.alpha.core.rules;
 import java.util.List;
 
 import at.ac.tuwien.kr.alpha.api.programs.Predicate;
-import at.ac.tuwien.kr.alpha.api.rules.NormalHead;
 import at.ac.tuwien.kr.alpha.api.rules.Rule;
+import at.ac.tuwien.kr.alpha.api.rules.heads.NormalHead;
 import at.ac.tuwien.kr.alpha.core.grounder.RuleGroundingInfo;
 
 public interface CompiledRule extends Rule<NormalHead> {
